refactor InvokeAIAppConfig
