+++ invokeai/app/services/config/__init__.py
@@ -1,0 +1,8 @@
+"""
+Init file for InvokeAI configure package
+"""
+
+from .invokeai_config import (
+    InvokeAIAppConfig,
+    get_invokeai_config,
+)
