+++ invokeai/backend/model_management/model_manager.py
@@ -341,7 +341,7 @@ class ModelManager(object):
         self.logger = logger
         self.cache = ModelCache(
             max_cache_size=max_cache_size,
-            max_vram_cache_size=self.app_config.max_vram_cache_size,
+            max_vram_cache_size=self.app_config.vram_cache_size,
             execution_device=device_type,
             precision=precision,
             sequential_offload=sequential_offload,
